requirements.txt
requests>=2.32.5
python-dotenv>=1.2.1
icalendar>=6.3.2
pytz>=2025.2